requirements.txt
bzip2
ca-certificates
expat
libcxx
libffi==3.4.4
ncurses==6.4
openssl==3.0.16
pydispatcher==2.0.7
python==3.12.9
readline==8.2
setuptools
sqlite==3.45.3
tk==8.6.14
tzdata
urrtde==2.7.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xz==5.6.4
zlib==1.2.13
